python/cases/movie_telegram_bot/requirements.txt
APScheduler
beautifulsoup4
certifi
python-dateutil
python-decouple
python-telegram-bot 
pytz
requests
six
soupsieve
tornado
urllib3
